+++ apps/api/src/modules/agents/service.ts
@@ -5,7 +5,7 @@ import { db } from "@hebo/db";
 import { agents } from "@hebo/db/schema/agents";
 import supportedModels from "@hebo/shared-data/supported-models.json";
 
-import type { AuditFields } from "~/middlewares/get-audit-fields";
+import type { AuditFields } from "~/middlewares/audit-fields";
 import { BranchService } from "~/modules/branches/service";
 import { createSlug } from "~/utils/create-slug";
